feat: 🎸 Move fetch order logic to OrderService
